backend/requirements.txt
fastapi==0.115.6
sqlmodel==0.0.31
uvicorn[standard]==0.34.0
pyjwt==2.9.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
bcrypt==4.2.0
